+++ build.ps1
@@ -2,6 +2,7 @@
               .\src\utilities.js,`
               .\src\custom.js,`
               .\src\feature-modules\chat.js,`
+              .\src\feature-modules\sessioning.js,`
               .\src\feature-modules\pm.js,`
               .\src\feature-modules\rng.js,`
               .\src\feature-modules\blocking.js,`