+++ build.sh
@@ -4,6 +4,7 @@ cat ./src/header.js \
     ./src/utilities.js \
     ./src/custom.js \
     ./src/feature-modules/chat.js \
+    ./src/feature-modules/session.js \
     ./src/feature-modules/pm.js \
     ./src/feature-modules/rng.js \
     ./src/feature-modules/blocking.js \